refactored pareto_optimizer code
